Intelligent character loading and some refactor
