+++ main.lua
@@ -1,6 +1,7 @@
 socket = require("socket")
 json = require("dkjson")
 require("util")
+require("consts")
 require("class")
 require("queue")
 require("globals")
@@ -13,7 +14,6 @@ require("input")
 require("network")
 require("puzzles")
 require("mainloop")
-require("consts")
 require("sound")
 require("timezones")
 require("gen_panels")
